C

Description

Store functors and Lenses as transformation between store functors

(Indexed) Store Comonad

data C s p x

(Indexed) Store Comonad

Constructors

C s (p -> x)

Manipulating Store comonads

pos :: C s p x -> s

peek :: C s p x -> p -> x

mapC :: (s -> s') -> (p' -> p) -> C s p x -> C s' p' x

compF :: C s p (C s' p' x) -> C (C s p s') (p, p') x

Composition of two C _ _ is isomorphic to another C _ _

decompF :: C (C s p s') (p, p') x -> C s p (C s' p' x)

Functor representatable as C

class Functor f => IsC s p (f :: Type -> Type) | f -> s p where

Minimal complete definition

Nothing

Methods

toC :: f x -> C s p x

default toC :: Coercible (f x) (C s p x) => f x -> C s p x

fromC :: C s p x -> f x

default fromC :: Coercible (C s p x) (f x) => C s p x -> f x

Lenses are natural transformations between Store comonads

newtype Lens s p s' p'

Constructors

Lens

Fields

unLens :: s -> (s', p' -> p)

(|>) :: Lens s p s' p' -> Lens s' p' s'' p'' -> Lens s p s'' p''

idLens :: Lens s p s p

toLens :: (forall x. C s p x -> C s' p' x) -> Lens s p s' p'

fromLens :: Lens s p s' p' -> C s p x -> C s' p' x

Encode finitary lenses into numbers

newtype LensCode

Lens for Finitary types, encoded as bunch of Integers.

Constructors

LensCode [(Integer, Integer)]

encodeLens :: (Finitary s, Finitary p, Finitary s', Finitary p') => Lens s p s' p' -> LensCode

Encode lens as LensCode

decodeLens :: (Finitary s, Finitary p, Finitary s', Finitary p') => LensCode -> Maybe (Lens s p s' p')

(Fallible) construction of Lens from LensCode

decodeLens' :: (Finitary s, Finitary p, Finitary s', Finitary p') => LensCode -> Lens s p s' p'

Fail-with-error version of decodeLens